src/ai-review-backend/requirements.txt
fastapi
uvicorn
openai
httpx
beautifulsoup4
python-dotenv
